build: Add optional typescript compile check (#111)
* build: Add optional typescript compile check

The typescript compiler can be run against javascript code and check for various potential issues based on its ability to infer typings.  This can also be supplemented with using JSDoc comments to specify types beyond what can be inferred by the typescript compiler.

Added npm script to execute compiler.  Eventually we may want to run as part of build/ci, but for now leaving it as optional to testing purposes.

* restore schema.service in dependencies.js for backward compatibility