fix: access checker config readonly
Access checker providers expect to be able to modify the config that is passed to them.  With the switch to node-config, changing the configs is not allowed.  Access checker now creates a copy of the access checker config to pass to the provider.  Also added some additional logging and converted the config access to use `config.get`

Had to update a few tests to use sinon to stub config.  These tests were previously directly mutating the config and not returning it properly to its original state.